std::vector<retaining_ptr<Technique>> Material::_techniques

+++ cocos/base/allocator/CCAllocatorStrategyGlobalSmallBlock.h
@@ -125,7 +125,7 @@ public:
     
     virtual ~AllocatorStrategyGlobalSmallBlock()
     {
-        for (int i = 0; i <= kMaxSmallBlockPower; ++i)
+        for (size_t i = 0; i <= kMaxSmallBlockPower; ++i)
             if (_smallBlockAllocators[i])
                 ccAllocatorGlobal.deallocate(_smallBlockAllocators[i]);
